+++ config/routes.rb
@@ -135,6 +135,8 @@ Rails.application.routes.draw do
       get 'last_closing_end', on: :collection
       get 'archive', action: 'download_archive', on: :member
     end
+    # export accounting data to csv or equivalent
+    post 'accounting/export' => 'accounting_exports#export'
 
     # i18n
     # regex allows using dots in URL for 'state'