+++ db/schema.rb
@@ -11,12 +11,12 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20190606074801) do
+ActiveRecord::Schema.define(version: 20190730085826) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
-  enable_extension "unaccent"
   enable_extension "pg_trgm"
+  enable_extension "unaccent"
 
   create_table "abuses", force: :cascade do |t|
     t.integer  "signaled_id"
@@ -202,10 +202,11 @@ ActiveRecord::Schema.define(version: 20190606074801) do
     t.string   "category"
     t.string   "export_type"
     t.string   "query"
-    t.datetime "created_at",  null: false
-    t.datetime "updated_at",  null: false
+    t.datetime "created_at",                   null: false
+    t.datetime "updated_at",                   null: false
     t.integer  "user_id"
     t.string   "key"
+    t.string   "extension",   default: "xlsx"
   end
 
   add_index "exports", ["user_id"], name: "index_exports_on_user_id", using: :btree